Fixed build warnings in MSVC

+++ include/daw/json/impl/daw_json_parse_string_quote.h
@@ -57,7 +57,7 @@ namespace daw::json::json_details::string_quote {
 			keep_going = not( q0 | q1 | q2 | q3 | q4 | q5 | q6 | q7 | s0 | s1 | s2 |
 			                  s3 | s4 | s5 | s6 | s7 );
 			keep_going = keep_going & static_cast<bool>( last - ( first + 8 ) >= 8 );
-			first += static_cast<int>( keep_going ) * 8;
+			first += static_cast<std::ptrdiff_t>( keep_going ) * 8;
 		}
 		first -= *( first - 1 ) == '\\' ? 1 : 0;
 	}
@@ -78,7 +78,7 @@ namespace daw::json::json_details::string_quote {
 			auto const s0 = test_at_byte<0U, '\\'>( buff );
 			keep_going = not( q0 | q1 | q2 | q3 | s0 | s1 | s2 | s3 );
 			keep_going = keep_going & static_cast<bool>( last - ( first + 4 ) >= 4 );
-			first += static_cast<int>( keep_going ) * 4;
+			first += static_cast<std::ptrdiff_t>( keep_going ) * 4;
 		}
 		first -= *( first - 1 ) == '\\' ? 1 : 0;
 	}
